Implement solution for the problem of finding the index with the minimum average difference
Implemented a solution to find the index with the minimum average difference in a given 0-indexed integer array. The algorithm calculates the prefix sum array to efficiently compute the sum of elements up to each index. It then iterates through each index, calculating the average difference and updating the minimum difference and index accordingly. Special handling is included for the last index to ensure correct calculation of the average difference. The solution is implemented in the `minimumAverageDifference` method of the `Solution` class.
Link: https://leetcode.com/problems/minimum-average-difference/